% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{area_circ}
\alias{area_circ}
\title{Area of Circle}
\usage{
area_circ(r)
}
\arguments{
\item{r}{value of side of radius}
}
\value{
Numerical value of area
}
\description{
This function finds the area of a circle
}
