% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{area_parallelogram}
\alias{area_parallelogram}
\title{Area of parallelogram}
\usage{
area_parallelogram(b, h)
}
\arguments{
\item{b, h}{base and height values of parallelogram}
}
\value{
Numerical value of area
}
\description{
This function finds the area of a parallelogram
}
