% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{area_sq}
\alias{area_sq}
\title{Area of Square}
\usage{
area_sq(x)
}
\arguments{
\item{x}{value of side of square}
}
\value{
Numerical value of area
}
\description{
This function finds the area of a square
}
