% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{area_tri}
\alias{area_tri}
\title{Area of triangle}
\usage{
area_tri(b, h)
}
\arguments{
\item{b, h}{base and height values of triangle}
}
\value{
Numerical value of area
}
\description{
This function finds the area of a triangle.
}
