% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{avgdist}
\alias{avgdist}
\title{Calculating Average Distance from the Mean}
\usage{
avgdist(input, na.rm = FALSE)
}
\arguments{
\item{input}{a vector of numbers}

\item{na.rm}{true or false for whether to remove NA's before performing the calculation}
}
\value{
A number
}
\description{
Calculates the average distance from the mean by converting each number in the
vector to an absolute value and then taking the arithmetic mean.
}
\examples{
a <- c(5,7,1,-3,-9)
avgdist(a)

b <- c(NA,4,7,-2,-12)
avgdist(b)

d <- c(NA,4,NA,-2,-12)
avgdist(d,na.rm=T)

}
