\name{basicMath-package}
\alias{basicMath-package}
\alias{basicMath}
\docType{package}
\title{
\packageTitle{basicMath}
}
\description{
\packageDescription{basicMath}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{basicMath}

Maintainer: \packageMaintainer{basicMath}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = basicMath)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{basicMath}}
%\section{Documentation Index}{\packageIndices{basicMath}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
