% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basictable.R
\name{basictable}
\alias{basictable}
\title{Constructing a Frequency Table}
\usage{
basictable(input, order = TRUE, na.rm = FALSE)
}
\arguments{
\item{input}{a vector of numbers}

\item{order}{whether the modes should be sorted, with true being the default}
}
\value{
A vector
}
\description{
Constructs a table which lists each unique entry in a vector and how many times
it appears in the vector
}
\examples{
a <- c(3,1,3,2,2,4,2,3,4)
basictable(a)

b <- c(6,9,1,6,2,2)
basictable(b, order = FALSE)

}
