% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{circumference}
\alias{circumference}
\title{Circumference of Circle}
\usage{
circumference(r)
}
\arguments{
\item{r}{value of side of radius}
}
\value{
Numerical value of area
}
\description{
This function finds the circumerence of a circle
}
