% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{divide}
\alias{divide}
\title{Divide Numbers}
\usage{
divide(x, y)
}
\arguments{
\item{x}{integer}

\item{y}{integer}
}
\value{
The division of two integers
}
\description{
This function divides two numbers
}
