% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{factorial}
\alias{factorial}
\title{Factorial}
\usage{
factorial(x)
}
\arguments{
\item{x}{integer}
}
\value{
The factorial of an integer
}
\description{
This function gets the factorial of x (i.e. x!)
}
