% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{geometricmean}
\alias{geometricmean}
\title{Calculating the Geometric Mean}
\usage{
geometricmean(input, na.rm = FALSE)
}
\arguments{
\item{input}{a vector of numbers}

\item{na.rm}{true or false for whether to remove NA's before performing the calculation}
}
\value{
A number
}
\description{
Calculates the geometric mean of a set of positive numbers. The function calculates
the product of every number in the vector and takes the nth root, where n is the number of
elements in the vector.
}
\examples{
a <- c(3,9,5,8)
geometricmean(a)

b <- c(NA,3,NA,7,9,2)
geometricmean(b)

d <- c(NA,3,NA,7,9,2)
geometricmean(d,na.rm=T)

}
