% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{hypotenuse}
\alias{hypotenuse}
\title{Hypotenuse}
\usage{
hypotenuse(x, y)
}
\arguments{
\item{x, y}{2 integer values (sides of a triangle)}
}
\value{
Numerical value of hypotenuse
}
\description{
This function finds the hypotenuse given two sides of a triangle.
}
