% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{is_triangle}
\alias{is_triangle}
\title{Is a Triangle}
\usage{
is_triangle(x, y, z)
}
\arguments{
\item{x, y, z}{3 integer values}
}
\value{
Boolean value if passes the triangle inequality
}
\description{
This function determines whether 3 values make a triangle.
}
