% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{med}
\alias{med}
\title{Calculating the Median}
\usage{
med(input)
}
\value{
A number
}
\description{
Calculates the median of a set of numbers. The function sorts the numbers in
the vector and selects the middle number (if the number of entries is odd) or takes the average
of the two middle numbers (if the number of entries is even)
}
\examples{
a <- c(5,9,4,1,11)
med(a)

b <- c(5,9,4,1,11,2)
med(b)

}
