% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{multiply}
\alias{multiply}
\title{Multiply Numbers}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{integer}

\item{y}{integer}
}
\value{
The multiplication of two integers
}
\description{
This function multiplies two numbers
}
