% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{sqr_root}
\alias{sqr_root}
\title{Square root}
\usage{
sqr_root(x)
}
\arguments{
\item{x}{integer}
}
\value{
The square root of an integer
}
\description{
This function returns the square root of an integer
}
