% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{SST}
\alias{SST}
\title{Calculate the Total Sum of Squares}
\usage{
SST(input)
}
\arguments{
\item{input}{a vector of numbers}
}
\value{
A number
}
\description{
Calculate the total sum of squares (SST) by subtracting the mean from each
number in the data set, squaring each re-centered observation, and then summing up the results.
}
\examples{
a <- c(4,12,8,7)
SST(a)

}
