% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{stdeviation}
\alias{stdeviation}
\title{Calculating the Standard Deviation}
\usage{
stdeviation(input, na.rm = FALSE, sample = TRUE)
}
\arguments{
\item{input}{a vector of numbers}

\item{na.rm}{true or false for removing NAs from the vector before calculations}

\item{sample}{true or false where true represents a sample standard deviation and false
represents a population standard deviation}
}
\value{
A number
}
\description{
Calculates the standard deviation. By selecting sample = TRUE, the user chooses
to calculate the sample standard deviation, where the sum of square is divided by n-1. By
selecting sample = FALSE, the calculation is treated as a population standard deviation, and
the sum of squares is divided by n.
}
\examples{
a <- c(2,4,6,8)
stdeviation(a)

b <- c(NA,2,4,NA,8)
stdeviation(b)

d <- c(NA,2,4,NA,8)
stdeviation(d,na.rm=T)

e <- c(NA,2,4,8,NA)
stdeviation(e,na.rm=T,sample=F)

}
