% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{stdevp}
\alias{stdevp}
\title{Calculating the Population Standard Deviation}
\usage{
stdevp(input)
}
\arguments{
\item{input}{a vector of numbers}
}
\value{
A number
}
\description{
If you have the entire population in your data set, instead of dividing the SST
by n-1, one divides by n before taking the square root.
}
\examples{
a <- c(2,4,6,8)
stdevp(a)

}
