% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{stdevs}
\alias{stdevs}
\title{Calculating the Sample Standard Deviation}
\usage{
stdevs(input)
}
\arguments{
\item{input}{a vector of numbers}
}
\value{
A number
}
\description{
Calculates the sample standard deviation, using n - 1 degrees of freedom.
}
\examples{
a <- c(2,4,6,8)
stdevs(a)

}
