% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicMath.R
\name{subtract}
\alias{subtract}
\title{Subtract Numbers}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{integer}

\item{y}{integer}
}
\value{
The difference of two integers
}
\description{
This function adds two numbers
}
