% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{arithmeticmean}
\alias{arithmeticmean}
\title{Calculating the Arithmetic Mean}
\usage{
arithmeticmean(input)
}
\value{
A number
}
\description{
Calculates the arithmetic mean of a set of numbers. The function adds up
the numbers in the vector and divides by the number of elements in the vector.
}
\examples{
a <- c(-4,9,1,-6)
arithmeticmean(a)

}
