% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{avgdist}
\alias{avgdist}
\title{Calculating Average Distance from the Mean}
\usage{
avgdist(input)
}
\arguments{
\item{input}{a vector of numbers}
}
\value{
A number
}
\description{
Calculates the average distance from the mean by converting each number in the
vector to an absolute value and then taking the arithmetic mean.
}
\examples{
a <- c(5,7,1,-3,-9)
avgdist(a)

}
