% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modes.R
\name{modes}
\alias{modes}
\title{Calculating the Mode}
\usage{
modes(input, order = TRUE)
}
\arguments{
\item{input}{a vector of numbers}

\item{order}{whether the modes should be sorted}
}
\value{
A vector of the most frequently occurring numbers
}
\description{
Determines which number(s) in the vector are the most frequently occurring.
}
\examples{
a <- c(1,2,3,3)
modes(a)

b <- c(1,2,2,2,3,3,3,4,4)
modes(b)

}
